[refactor] [move] FE playerCard 영역 - domain 별 구분을 위해 파일 이동 REF : #18
