+++ web/src/app/router/routes/AdminRoutes.jsx
@@ -10,7 +10,7 @@ const AdminEventPage = lazy(() => import("@/domains/events/feature/admin/pages/A
 const AdminNoticeManagePage = lazy(() => import("@/domains/notices/feature/components/admin/AdminNoticeManagePage.jsx"));
 const AdminCommunityPage = lazy(() => import("@/domains/community/page/admin/AdminCommunityPage.jsx"));
 const AdminCouponListPage = lazy(() => import("@/domains/coupons/feature/admin/pages/AdminCouponListPage.jsx"));
-const AdminPlayerPage = lazy(() => import("@/domains/admin/pages/player/pages/AdminPlayerPage.jsx"));
+const AdminPlayerPage = lazy(() => import("@/domains/playerCard/feature/admin/pages/AdminPlayerPage.jsx"));
 
 export const AdminRoutes = [
   {
