+++ web/src/domains/playerCard/config/PlayerTableConfig.js
