+++ web/src/domains/playerCard/feature/admin/components/table/PlayerTable.module.scss
