+++ web/src/domains/playerCard/feature/admin/pages/AdminPlayerPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AdminPlayerTable from "@/domains/admin/pages/player/components/AdminPlayerTable.jsx";
+import AdminPlayerTable from "@/domains/playerCard/feature/admin/components/AdminPlayerTable.jsx";
 
 const AdminPlayerPage = () => {
   return <AdminPlayerTable />
